Add a FIXME for an RStudio-specific issue, and a TODO for guidance.
The other RAs might benefit from this. For now, I consider this low priority
since it does not seem to affect the application instance running on
shinyapps.io, and it does not appear when using the application in ESS.

ESS does not offer a facility/keybind/button to "run [a Shiny application]", and
it uses the normal mode of evaluation in R: simply executing each statement in
turn. I suspect this is the issue in RStudio Desktop that @ashokkrish has found.
The issue should have been present *before* the changes made to this file as
well, given that the order of evaluation is the same, for the most part.