Fix #6 and make LaTeX rendering dynamic
The tabsetPanel (the main/output panel) containing the four tabs now has proper
visibility upon application startup, after running a simulation of a model, and after
changing to a different model; this fixes #6.

The mathematical model tab contents (LaTeX) is now rendered dynamically with
renderUI. Conveniently, this is the route I need to take to resolve a different
issue (which cannot be named in this message or GitHub may automatically resolve
that issue when it hasn't had a status change in reality).