A very small step towards fixing sixty.
This will help Khanh resolve the issues with input validation by making the
logic here correct and sane. Before, it was not correct and could not be
expected to result in the correct behaviour. Now it will result in the intended
behaviour, however it does not fix the issue. *Validator objects are still not
found in any environment, let alone the global one where they are assigned to
and where they are attempted to be retrieved from. This is strange behaviour;
Khanh is working on resolving this.