Progress on #52 and #43
made progress towards #52 but there is a bug in the ui that un-aligns control elements when a custom order is created using 'dom' in DT::dataTable options. Secondly, Only the first compartment in the subplots seems to get its compartment title - once this is fixed I will add subplot functionality to the other models.