Implement the model configuration LaTeX re-rendering.
- The incorrect in-line delimiter was used previously. The correct delimiter is
  now used. See https://docs.mathjax.org/en/latest/basic/mathematics.html.
- Model configuration LaTeX is now rendered whenever the model changes, by using
uiOutput rendering instead of JavaScript to trigger the rendering manually (fix #5).
- The position of updateNumericInputs is moved to after the rendering of the
beta and gamma widgets to assure that their target is rendered before their
numeric value is modified.
- The JavaScript that conditionally modifies the padding of the modelSelect
widget, with respect to the selection, is now working (fix #4).
- The way the input object was accessed in JavaScript in the previous iteration
of `modelSelect.js` and its minified copy broke reactivity; the object is now
accesse correctly in that script, so reactivity is no longer broken (fix #69).