Candidate fix №1 for unified model solver issues
The reactive values of input widgets passed as arguments to the model solver
function are now only those which are visible. JavaScript was used to filter
these values and produce a new reactive value which is then used in
`doCall.default()`. This should fix #82.

These changes are rebased onto 41380dfd935407f4eca640681895291880107fb2.