Update commentary; rewrite createSubPlotsUI function; update spreadsheet.
- Update commentary in sever: removing mention of a deleted function from
sectioning comments used by RStudio, and adding description of why two calls to
shinyjs::disable() are made (disable inputs and navigation related to
unimplemented features); rephrase sentence about recycling inputId. Remove
unnecessary NOTE; remove unnecessary and outdated comments; removed poor descriptions
of some lines of code; rephrased an important warning ("WARN DONT") to make it
clearer what not to do; rephrased commentary on the modelSelect observable;
removed mention of responsibility of a JavaScript file that no longer exists/was outdated.

- Wrapped long strings in gsubs to break the lines using raw strings and then
rejoin them when they're used.

- Rewrote `createSubPlotsUI()` function to use iwalk and make the function
shorter and easier to read.

- Updated the default values in the spreadsheet and the default values for
muBirth and muDeath in the UI to fix an on-start issue where when vital dynamcis
is enabled the rates are zero by default (which doesn't make sense to me, and
wont' make sense to end-users).