Introduce a reactive expression returning the visible inputs
The modelSolver is now called with doCall.default using a list of arguments,
named values, which are taken from the user interface and are only those which
are truly visible. Previously, these mistakenly included some invisible inputs,
such as delta, due to an incorrect use of the extract operator and
concatenation (fix #82). It now works as intended. The fix has revealed a new
issue (#92), or introduced it somehow.