Unify solvers, "simplifying" codebase. Fix renderLaTeX call.
The call to renderLaTeX was broken because the inputs() reactive was obsoleted;
there is now a visibleInputs() reactive. While fixing this function call I
cleaned up the function in model_LaTeX.R to make it more readable and use the same
style as other functions in the application.

The unified solver function is called exposuRe, and it uses a unified function
of differential equations called "differentialEquations."

The separate solvers are each repaired, properly checking the available
variables. Prior to this commit, the following line caused the issue.

if (delta == 0 && D != 0) stop("D (dead) must be zero when delta is zero.")

The comparison "D != 0" was not semantically correct when the dead compartment
was disabled. The comparison failed because the name "D" pointed to stats::D,
because "D" := dead was removed from the "variables" vector prior to calling
lsoda. While dead = 0 when the dead compartment is disabled, the symbol "D" does
not apply to that compartment, because prior to this commit the dead compartment
was removed from the "variables" vector. That was a design mistake, surely. This
repaired version does not do anything like that. All variables remain, because a
variable which is always zero is perfectly acceptalbe to lsoda.