Obsolte separate solvers for SI-type and SEI-type models.
Realized the cause of #91 was the (unnecessary and optional) inclusion of
trueMassAction in the list returned from the differentialEquations function (fix

Made the differentialEquations function local to the unified solver: exposuRe.

Changed the way lists and vectors are joined together in exposuRe, ensuring:
vectors remain vectors and are not coerced to lists when undesirable; lists are
appended to correctly; coersion is allowed to take place when necessary.

Added protections for some edge cases that I did not anticipate in the first
edition of the unified solver.

Resolved an ephemeral regression (having no tracking issue) which caused the
number of returned derivatives to be one greater than the number of state
variables when running SEIRS or SIRS models. This was caused by an assignment of
the modified form dS (when loss of immunity is enabled) to a component of the
compartment local variable, which was erroneous.

Included more commentary and named vector elements and list components
descriptively, which aided in debugging this edition and making it a viable
product.

Reworded some existing commentary.