Fully deprecate the separate model solvers.
The reactive expression modellingFunctions no longer gets the name of a
model-specific solver, because there are no model-specific solvers.

The lines related to testing of the new, unified model solver are removed. The
new, unified solver is on par for functionality, producing identical results in
all cases.

Extraneous commentary related to the model solver is removed from server.R.