Revert "Move createSubPlotsUI to UI.R and rename it"
I forgot that because the function is used in server it must be
available in server. It is not available because UI.R is sourced once,
and its result passed to the server function upon initialization,
hence the functions defined in UI.R which are not used in UI.R to make
the final result of the script are unavailable.

This commit moves the function back into server.R, and undoes the rename.

This reverts commit 51b9edb17c1577dfb0210fea8a49c72375e46fd6.