Monolithic changes across application; see details.
The changes include the following, in summary: renamed files in R/ subfolder to
make them "funnie.R" (that's important, indeed); renamed a misspelled file in
R/; instant input validation is more robust and also includes more flexibility;
RStudio Desktop's R console is able to run the app again; partial application
reset works as requested; default values are fixed and as requested; beta and
gamma reactivity insofar as default values is fixed.

In detail, the changes are as follows.

The cause of the beta and gamma widget's default values not being applied when
SIR or other models were first selected hasn't been fully determined, but the
issue has been resolved nonetheless (fix #90). As far as I'm aware, only the SIR
model selection was actually affected; I didnt' see it with others, but it's
possible they were also affected.

The default values Ashok requested for an SIR model type and an SEIR model type
have been adjusted (fix #89).

RStudio Desktop can now run the application in its R console again; why this is
the case is unknown to me, and it feels like an impenetrable issue. It is no
longer necessary to change the option to "In Background Job" in the Run App
dropdown menu.

The solution to #82 has been made more robust, and a new feature has been
introduced. The model selection is now matched against a ruleset with a regular
expression, allowing rules that govern input widgets in multiple models to be
specified a single time with a ruleset mathing those multiple models. For
example, all models with a Dead compartment (models with fatal disease) now have
a single ruleset for the Dead initial state variable input widget and the
delta (fatality rate) input widget. Only models with very specific rules will
need to have their own ruleset, and this is unlikely until we discover that this
is the case in the future (if it ever is). Regardless, the feature exists, is
tested, and makes the rules more readable and less repetitive.