Make outputPanel contents reactive to all relevant inputs
The visibleInputs now cause the outputPanel to react to changes and immediately
update all dependent outputs (the simulation and its plots, LaTeX, etc.). This
required a days worth of refactoring, working from 1:15PM to 6:15PM on the
changes, and reviewing new concepts in Bootstrap which will supplement the
visual exerpience thereafter and documenting changes in the last fifteen
minutes, ending at 7PM.

The assumption of a closed population is now validated in all cases, using
shinyFeedback (fix #95). The dead compartment is not considered in the summation
check, as it never impacts the number of active infections (theres a difference
between the dead compartment and the infectious compartment; those in the
infectious compartment could be corpses, but they're generally just considered
infectious).

The new reactivity continuously re-renders the UI of the outputPanel, so the
JavaScript which disabled the Basic Reproduction Number (R₀) tabpanel no longer
works, because the JavaScript would also need to watch for this change and its
not worth it. In the interim, while the BRN tabpanel has no content it is simply
commented out.

Functions in server.R which don't directly depend on reactive values are now
moved to global.R to make server.R shorter. This has the added benefit of
increasing the initial connection speed of new clients when they connect to the
Shiny application.

The waiter application is included, however autoWatier() has not had the
intended effect and the waitstaff will need to be programmed directly.

The server is updated to render the UI for the outputPanel, and the UI is
updated likewise. Functions are sometimes inlined, and general improvements are
made in this effort that introduces reactivity.