Do not reset input widgets when model options change
This commit removes the observeEvent which was used to reset application
variables when the model options changed (fix #99). This influences the design
decisions related to issue #49, @ashokkrish; removing the observeEvent which
updates the numeric input widgets to values from the defaultInputValues
dataframe and spreadsheet removes some "wizardry" from the application. We will
need to remove all of the rows from the spreadsheet which are predicated upon
the combinations of these selections. This change also means we will need to
change the filtering of the dataframe in the defaults() reactive, since that no
longer considers these model options.