Update Author tab in UI
A new function, developer(), is written to generate the HTML used in the author
tab to include an individual developer; there are improvements that could be
made to this, but it is better than what was there before.

Ashok Krishnamurthy's (@ashokkrish) address information is updated in style to
make it appear more professional (in a web user's perception).