Add stochastic rerun button to conditional stochastic wellPanel
The button used to rerun a stochastic simulation appears next to the number of
replicates in the "stochastic" conditional wellPanel.

Additionally, autoWaiter is commented out of the UI because the waiter library
isn't used right now; autoWaiter wasn't working.

Further, the reset button is moved out of a div and renamed