Fix circular reactivity caused by extraneous input$plotly_* objects
Plotly "pollutes" input, surprisingly to me, and this caused a circular
reactivity issue. Because plotly edits its own objects underneath input this
causes visibleInputs to be invalidated downstream (it depends on the global
input object). This caused the modelling process to be re-run because
visibleInputs changed, and this restarted the cycle. This is now corrected
(fix #101).

Include automatic issue closures for things I fixed in previous commits and
forgot to include (I don't want to amend the wording of all the commits):

visibleInputs is debounced (fix #101).

defaultInputValues is updated with the values requested by @ashokkrish (fix #89).

The styling of inputs underneath the #modelConfiguration div are now styled
nicely, grouped into variables and parameters (with some titling; fix #72).

A new button for re-running stochastic simulations is included (fix #102).

The debounce delay for the visibleInputs() reactive is increased to 1.25
seconds, which seems to fix most issues with reactive value propagation (fix #98).

That seems to be it!