Fix default value updating and conditionally allow stochastic models
Stochastic models are enabled for SIR models only, and the dynamic updating of
default values is fixed.

Further rows of default values are required for the application to work
perfectly; Ashok will be the best person to make this commit in the morning,
since he'll know exactly what values might be useful. We can only guess and
throw in random/kludge values or copy values from other simulations which might
not make theoretical sense.