Implement the conditional inclusion of literature-based defaults
Fully prevent the selection of a uniform distribution function when the
simulation will be stochastic and total mass action is selected (fix #130).

Resolve the issues related to the now-implemented UI and server functionality to
use literature-based model defaults. The list of literature-based defaults is
only added to the model selection dropdown when there is at least one
literature-based model; otherwise only the application-default compartmental
models are shown in the model selection dropdown (fix #11).