Resolve multiple bugs present in release 2.0-beta
Unfortunately, release 2.0-beta had several glaring issues present which were
not detected through manual testing. This is more reason that we should
definitely adopt an automated testing framework for Episim, SpatialEpisim, and
any other Shiny applications the lab works on.

The renamed input "freeze" had a change in the implied meaning of its boolean
value; one of the relevant lines that checks this boolean value was not updated
to relfect the mentioned change. The condition using the boolean value is
updated in this commit so that when the user unchecks the Freeze button (they
thaw the input widgets) the input widget values will be updated, and this also
causes a reactive chain of events to update the plot in this and other
cases (fix #135). The total-mass action simulation for SIR or other simulations
is now updated correctly, and the plot appears correct.

A function which was removed and is no longer relevant to the application had a
function call to it remaining in the source code when plot settings were
removed; this line calling the function is now removed (fix #134).

The output summary tab panel is now only displayed for deterministic
simulations; to speed up the application, the dependency of this output panel, a
datatable, is only created when the simulation is not stochastic.

The bold header reading "COMPARTMENTAL MODEL" is now capitalized again.

A new `req`, which also uses the `cancelOutput = "progress"` argument, was introduced
to further protect the main output panel from displaying error messages related
to #98. Why the line doesn't have the intended effect is unknown to me, so it
remains in a comment for now, because it doesn't solve the issue it was
introduced for.

The main panel (all of the tabPanels) is no longer hidden and shown when the
visible inputs change (something that was introduced when I was playing with
using a Run button to avoid reactive value propagration issues (see #98). A run
button breaks instant input validation in some ways (see #131), so that was
removed. These two lines were removed because they were extraneous, even though
they didn't directly cause any visual errors or reactive value propagation
bugs (as far as I could see).

- shinyjs::hide(selector = ".col-sm-8")
- shinyjs::show(selector = ".col-sm-8")

The change from XLSX to XLS in an earlier commit is now reflected by updating
removing the backup ZIP of an older version of the XLSX file from the
repository, and removing the XLSX from the repository.

The XLS file is updated by removing applied filters, so when the spreadsheet is
opened by a contributor they will see all observations immediately; in the
future we should be careful not to save the spreadsheet with filters applied (if
used). Note that the filters used in a spreadsheet application should not make
the values unavailble to our application, but it is better to be safe than sorry.