Include new react-chart-editor in Episim
Unfortunately there is a glaring CSS issue in the `div.fold` elements, but there's
nothing I can seem to do about it without much more research. The temporary
solution is to open the developer tools window in Chrome, Firefox, Safari, etc.
and remove the `box-sizing = 'border-box'` property from the blue fold bars that
make up a significant amount of the chart editor ui. I've tried to do this
automatically, spending an hour on this, but with no luck.