Update the git submodule for reactcharteditor.
Toby rebuilt the JSX library so that Plotly plots configuration does not
include `editable = TRUE`, i.e. `editable = FALSE` is now passed to the
`config` argument in the Plotly function call.

This is done because we do not want users to click plot axes titles or
the plot title to change it. We don't want them to be able to do that
because the change is not propagated from the embedded Plotly chart in
the editor to the rest of the editor and then back to the Model tab. The
user needs to first click the panel of the editor somewhere to make the
changes recognized, and then the update propagates to the Model tab.
We don't want this to be an issue raised later, so we're prohibiting
the feature for now, until the plotly/react-chart-editor people fix it
themselves.